# backend/handlers/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PAIMOS API",
    "description": "PAIMOS — Professional & Personal AI Project OS.\n\nThis is the **public, stable scriptable contract** for PAIMOS — the canonical resource surface used by the SPA, the paimos CLI, and custom integrations: auth, users, projects, issues, comments, time entries, attachments, documents, incidents, knowledge, and GDPR ops. It is intentionally a curated subset. Internal one-off admin tooling (CSV/Jira/mite imports, dev test reports, branding writes, SSO/TOTP management, AI operations, portal internals) is deliberately omitted and is NOT part of the stability contract — those endpoints exist but may change without notice. Authentication is either a session cookie (browser) or a `paimos_*` bearer API key (machine-to-machine). Coverage policy & contribution rule: PAI-294.",
    "version": "1.0.0",
    "license": {
      "name": "AGPL-3.0-only",
      "identifier": "AGPL-3.0-only"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin (recommended)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "Issued by POST /api/auth/login. Browser SPA. CSRF-protected via X-CSRF-Token header for mutations (PAI-113)."
      },
      "bearerAPIKey": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "paimos_*",
        "description": "Long-lived API key, prefix configurable via BRAND_API_KEY_PREFIX. Created via /api/auth/api-keys."
      }
    },
    "schemas": {
...
      "CommentCreateRequest": {
        "type": "object",
        "required": ["body"],
        "properties": {
          "body": {"type": "string", "minLength": 1},
          "visibility": {"type": "string", "enum": ["internal", "external"], "default": "internal"},
          "client_request_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "description": "Accepted only when visibility is internal. The database identity, rather than the generic response cache, is authoritative."
          }
        }
      },
...
      "Comment": {
        "type": "object",
        "required": [
          "id",
          "issue_id",
          "body",
          "created_at"
        ],
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "issue_id": {
            "type": "integer",
            "format": "int64"
          },
          "author_id": {
            "type": "integer",
            "format": "int64",
            "nullable": true
          },
          "body": {
            "type": "string"
          },
          "visibility": {
            "type": "string",
            "enum": ["internal", "external"]
          },
          "client_request_id": {
            "type": "string",
            "nullable": true,
            "minLength": 1,
            "maxLength": 128,
            "pattern": "^[A-Za-z0-9._:-]+$",
            "description": "Optional exact-once identity scoped to the authenticated author. Legal only for internal comments; an exact replay returns the original comment and different issue/body reuse returns 409."
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "AgentModeProblem": {
        "type": "object",
        "additionalProperties": false,
        "required": ["type", "title", "status", "detail", "code", "error"],
        "properties": {
          "type": {"type": "string", "format": "uri"},
          "title": {"type": "string"},
          "status": {"type": "integer"},
          "detail": {"type": "string"},
          "instance": {"type": "string"},
          "code": {"type": "string"},
          "field": {"type": "string"},
          "valid_values": {"type": "array", "items": {"type": "string"}},
          "request_id": {"type": "string"},
          "error": {"type": "string"}
        }
      },
...
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    },
    {
      "bearerAPIKey": []
    }
  ],
  "paths": {
...
    "/api/issues/{id}/comments": {
      "get": {
        "tags": [
          "comments"
        ],
        "summary": "List comments on issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Comments",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Comment"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "comments"
        ],
        "summary": "Create comment on issue",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/CommentCreateRequest"}}}
        },
        "responses": {
          "200": {
            "description": "Exact replay of an existing client_request_id",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Comment"}}}
          },
          "201": {
            "description": "Created",
            "content": {"application/json": {"schema": {"$ref": "#/components/schemas/Comment"}}}
          },
          "400": {
            "description": "Invalid body, visibility, or client_request_id"
          },
          "409": {
            "description": "client_request_id was already used by this author for a different issue or body",
            "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/AgentModeProblem"}}}
          }
        }
      }
    },
...
  }
}